+++ packages/containers-shared/src/utils.ts
@@ -50,15 +50,18 @@ export const runDockerCmdWithOutput = async (
 };
 
 /** throws when docker is not installed */
-export const verifyDockerInstalled = async (dockerPath: string) => {
+export const verifyDockerInstalled = async (
+	dockerPath: string,
+	isDev = true
+) => {
 	try {
 		await runDockerCmd(dockerPath, ["info"], ["inherit", "pipe", "pipe"]);
 	} catch {
-		// We assume this command is unlikely to fail for reasons other than the Docker CLI not being installed or not being in the PATH.
+		// We assume this command is unlikely to fail for reasons other than the Docker daemon not running, or the Docker CLI not being installed or in the PATH.
 		throw new Error(
-			`The Docker CLI does not appear to installed. Please ensure that the Docker CLI is installed. You can specify an executable with the environment variable WRANGLER_DOCKER_BIN.\n` +
-				`Other container tooling that is compatible with the Docker CLI may work, but is not yet guaranteed to do so.\n` +
-				`To suppress this error if you do not intend on triggering any container instances, set dev.enable_containers to false in your Wrangler config or passing in --enable-containers=false.`
+			`The Docker CLI could not be launched. Please ensure that the Docker CLI is installed and the daemon is running.\n` +
+				`Other container tooling that is compatible with the Docker CLI and engine may work, but is not yet guaranteed to do so. You can specify an executable with the environment variable WRANGLER_DOCKER_BIN and a socket with WRANGLER_DOCKER_HOST.` +
+				`${isDev ? "\nTo suppress this error if you do not intend on triggering any container instances, set dev.enable_containers to false in your Wrangler config or passing in --enable-containers=false." : ""}`
 		);
 	}
 };