+++ packages/wrangler/src/cloudchamber/deploy.ts
@@ -1,6 +1,7 @@
 import { isDockerfile } from "@cloudflare/containers-shared";
 import { type Config } from "../config";
 import { type ContainerApp } from "../config/environment";
+import { containersScope } from "../containers";
 import { getDockerPath } from "../environment-variables/misc-variables";
 import { UserError } from "../errors";
 import { isNonInteractiveOrCI } from "../is-interactive";
@@ -17,14 +18,17 @@ export async function maybeBuildContainer(
 	imageTag: string,
 	dryRun: boolean,
 	pathToDocker: string
-) {
+): Promise<{ image: string; pushed: boolean }> {
 	try {
 		if (
 			!isDockerfile(
-				containerConfig.image ?? containerConfig.configuration.image
+				containerConfig.image ?? containerConfig.configuration?.image
 			)
 		) {
-			return containerConfig.image ?? containerConfig.configuration.image;
+			return {
+				image: containerConfig.image ?? containerConfig.configuration?.image,
+				pushed: false,
+			};
 		}
 	} catch (err) {
 		if (err instanceof Error) {
@@ -36,13 +40,13 @@ export async function maybeBuildContainer(
 
 	const options = getBuildArguments(containerConfig, imageTag);
 	logger.log("Building image", options.tag);
-	const tag = await buildAndMaybePush(
+	const buildResult = await buildAndMaybePush(
 		options,
 		pathToDocker,
 		!dryRun,
 		containerConfig
 	);
-	return tag;
+	return buildResult;
 }
 
 export type DeployContainersArgs = {
@@ -62,7 +66,11 @@ export async function deployContainers(
 	}
 
 	if (!dryRun) {
-		await fillOpenAPIConfiguration(config, isNonInteractiveOrCI());
+		await fillOpenAPIConfiguration(
+			config,
+			isNonInteractiveOrCI(),
+			containersScope
+		);
 	}
 	const pathToDocker = getDockerPath();
 	for (const container of config.containers) {
@@ -105,17 +113,25 @@ export async function deployContainers(
 			],
 		};
 
-		const image = await maybeBuildContainer(
+		const buildResult = await maybeBuildContainer(
 			container,
 			versionId,
 			dryRun,
 			pathToDocker
 		);
+		container.configuration ??= {};
+		container.configuration.image = buildResult.image;
+		container.image = buildResult.image;
 
-		container.configuration.image = image;
-		container.image = image;
-
-		await apply({ skipDefaults: false, json: true, env }, configuration);
+		await apply(
+			{
+				skipDefaults: false,
+				json: true,
+				env,
+				imageUpdateRequired: buildResult.pushed,
+			},
+			configuration
+		);
 	}
 }
 
@@ -127,7 +143,7 @@ export function getBuildArguments(
 	container: ContainerApp,
 	idForImageTag: string
 ): BuildArgs {
-	const imageRef = container.image ?? container.configuration.image;
+	const imageRef = container.image ?? container.configuration?.image;
 	const imageTag = container.name + ":" + idForImageTag.split("-")[0];
 
 	return {